+++ module/data/item/trick.d.ts
@@ -1,0 +1,12 @@
+import { DotsData } from "@models/fields/dots";
+
+export {};
+
+declare module "./trick.mjs" {
+	interface Trick {
+		type: keyof typeof curseborne.config.trickTypes;
+		cost: DotsData & { type: keyof typeof curseborne.config.trickCostTypes };
+		/** Whether the trick can be purchased multiple times in the same action. */
+		multiple: boolean;
+	}
+}